Rename test file to match file under test
